+++ src/animatable/shadow.rs
@@ -2,7 +2,10 @@
 
 use iced::Shadow;
 
-use crate::animatable::{color::lerp_color_raw, geometry::lerp_vector_raw, lerp_f32_raw};
+use crate::{
+    animatable::{geometry::lerp_vector_raw, lerp_f32_raw},
+    color::AnimColor,
+};
 
 use super::{Animatable, InterpolationProgress, interpolate_with_progress};
 
@@ -12,7 +15,7 @@ impl Animatable for Shadow {
             let progress = progress.value();
 
             Self {
-                color: lerp_color_raw(from.color, to.color, progress),
+                color: AnimColor::interpolate(from.color.into(), to.color.into(), progress).into(),
                 offset: lerp_vector_raw(from.offset, to.offset, progress),
                 blur_radius: lerp_f32_raw(from.blur_radius, to.blur_radius, progress),
             }